Privileged vs unprivileged script execution problems are sorted

+++ setup/rails.sh
@@ -11,18 +11,27 @@ echo =       Installing rails
 echo =
 echo ===================================
 
-# Environment Vars
-export PATH="$HOME/.rbenv/bin:${PATH}"
-export PATH="$HOME/.rbenv/shims:${PATH}"
-export RBENV_SHELL=bash
+provisioned=/home/vagrant/provisions/rails
+if [ -f "$provisioned" ]
+then
+  echo "rails has already been provisioned"
+else
+  # Environment Vars
+  export PATH="$HOME/.rbenv/bin:${PATH}"
+  export PATH="$HOME/.rbenv/shims:${PATH}"
+  export RBENV_SHELL=bash
 
-# Set ruby version
-rbenv local 2.2.0
+  # Set ruby version
+  rbenv local 2.2.0
 
-# Rails
-echo  Installing rails
-gem install rails
-rbenv rehash
+  # Rails
+  echo  Installing rails
+  gem install rails
+  rbenv rehash
+
+  # Set provisioned flag
+  touch $provisioned
+fi
 
 echo ---
 echo rails installation is complete!
